[추가] 그래프 구현
선택한 센서와 공기질 데이터 종류에 대한 데이터를 전달해주는 api 구현

+++ src/main/java/com/awexomeray/TwoParkHanJungLim/dto/mainDto/AirDataDto.java
@@ -10,14 +10,14 @@ public class AirDataDto {
 
     private Double humi;
 
-    private Integer co2;
+    private Double co2;
 
-    private Integer tvoc;
+    private Double tvoc;
 
-    private Integer pm01;
+    private Double pm01;
 
-    private Integer pm25;
+    private Double pm25;
 
-    private Integer pm10;
+    private Double pm10;
 
 }
